demo_recordia/requirements.txt
streamlit>=1.53.0
openai>=2.15.0
pandas>=2.3.0
matplotlib>=3.10.0
psycopg2-binary>=2.9.9
